docking and undocking boats, clean up other functions
note: ready for formal testing with postman

+++ main.py
@@ -1,13 +1,12 @@
 import json
+import datetime
 from google.appengine.ext import ndb
 import webapp2
 
-# EXAMPLE:
-class Fish(ndb.Model):
-    name = ndb.StringProperty(required=True)
-    ph_min = ndb.IntegerProperty()
-    ph_max = ndb.IntegerProperty(repeated=True) #repeated allows lists of properties (makes it an array)
 
+# +--------------------------------------+
+# |          Modules Used                |
+# +--------------------------------------+
 
 class Boat(ndb.Model):
     id = ndb.StringProperty() # for this I should use the primary key
@@ -17,44 +16,17 @@ class Boat(ndb.Model):
     at_sea = ndb.BooleanProperty(default=True)
 
 
-class Departure_Ticket(ndb.Model):
-    departure_date = ndb.DateProperty()
-    departed_boat = ndb.KeyProperty()
-
-
 class Slip(ndb.Model):
     id = ndb.StringProperty() # for this I should use the primary key
     number = ndb.IntegerProperty(required=True)
-    current_boat = ndb.KeyProperty()
+    current_boat = ndb.StringProperty()
     arrival_date = ndb.StringProperty()
-    # departure_history = 
 
 
 
-# EXAMPLE:
-class FishHandler(webapp2.RequestHandler):
-    def post(self):
 
-        parent_key = ndb.Key(Fish, "parent_fish")
 
-        fish_data = json.loads(self.request.body)
-        new_fish = Fish(name=fish_data['name'], parent=parent_key)
-        new_fish.put()
-        fish_dict = new_fish.to_dict()
-        fish_dict['self'] = '/fish/' + new_fish.key.urlsafe()
-        self.response.write("here's that data: " + json.dumps(fish_dict))
 
-    def get(self, id=None):
-        if id:
-            f = ndb.Key(urlsafe=id).get()
-
-            # aside: how to update a property on a model:
-            f.ph_max = [100, 50]    # change property here
-            f.put()                 # update model (final step)
-
-            f_d = f.to_dict()
-            f_d['self'] = "/fish/" + id
-            self.response.write(json.dumps(f_d))
 
 
 
@@ -94,12 +66,15 @@ class BoatHandler(webapp2.RequestHandler):
         boat_dict['self'] = '/boat/' + new_boat.key.urlsafe()
         self.response.write("Here's the boat: " + json.dumps(boat_dict))
 
+
     # GET an existing BOAT
     def get(self, id=None):
         if id:
-            b = ndb.Key(urlsafe=id).get()
+            idInt = int(id)
+            b = Boat.get_by_id(idInt)
             b_d = b.to_dict()
             b_d['self'] = "/boat/" + id
+            self.response.headers["Content-Type"] = "application/json"
             self.response.write(json.dumps(b_d))
 
         # display ALL boats:
@@ -111,16 +86,168 @@ class BoatHandler(webapp2.RequestHandler):
             # put all slips in a list for nice printing:
             boat_list = []
             for boat in allBoats:
+                boatID = boat.id
                 boat = boat.to_dict()
+                boat['self'] = "/boat/" + boatID
                 boat_list.append(boat)
             self.response.headers["Content-Type"] = "application/json"
             self.response.write(json.dumps(boat_list))
-            self.response.write("end of list")
 
 
     # delete a boat
     def delete(self, id=None):
-        self.response.write("how dare you!")
+
+        if id:
+            # look up slip with given id
+            b = Boat.get_by_id(int(id))
+
+            if b is None:
+                self.response.set_status(404)
+                self.response.write("Boat with given id could not be found")
+
+            else:
+                # Boat at sea. Just delete boat:
+                if (b.at_sea == True):
+                    b.key.delete()
+                    self.response.set_status(200)
+                    self.response.write("Boat with given id was removed successfully")
+
+                # boat is docked. Remove reference to boat first:
+                else:
+                    # query for slip that has this boat attached
+                    q = Slip.query().filter(Slip.current_boat == b.id)
+                    slips = q.fetch()
+                    
+
+                    # 'undock' boat from slip(s) before remove:
+                    for slip in slips:
+                        slip.current_boat = None
+                        slip.arrival_date = None
+                        slip.put()
+
+                    # finally delete boat:
+                    b.key.delete()
+                    self.response.set_status(200)
+                    self.response.write("Boat with given id was removed successfully")
+
+        # no id supplied
+        else:
+            self.response.set_status(404)
+            self.response.write("You must supply an id for the boat to be removed")
+
+
+    # Modify a boat
+    def patch(self, id=None):
+        if id:
+            # look up boat with given id
+            b = Boat.get_by_id(int(id))
+            if (b != None):
+
+                # used for docking
+                slipNum = 0     # start with 0
+
+                # get current date:
+                now = datetime.datetime.now().date()
+
+                # get body data:
+                boat_data = json.loads(self.request.body)
+
+                # check for provided date:
+                if boat_data.has_key("date"):
+                    date = boat_data['date']
+                else:
+                    date = str(now)
+
+                status = True   # default value (for scope reasons)
+                # check for provided boat status:
+                if boat_data.has_key("at_sea"):
+                    statusStr = boat_data['at_sea']
+                    status = statusStr.lower() in ("yes", "true", "t", "1")
+                else:
+                    statusStr = b.at_sea #previus value
+                    status = statusStr.lower() in ("yes", "true", "t", "1")
+
+
+                # UN-dock boat:
+                if status:
+                    # query for any slips that have this boat docked (ALL in case of error)
+                    q = Slip.query().filter(Slip.current_boat == b.id)
+                    slips = q.fetch()
+
+                    # 'undock' boat from slip(s):
+                    for slip in slips:
+                        slip.current_boat = None
+                        slip.arrival_date = None
+                        slip.put()
+
+                    # put boat out to sea:
+                    b.at_sea = True
+                    b.put()
+                    self.response.set_status(200)
+                    self.response.write("Boat successfully undocked!")
+
+                # DOCK the boat:
+                else:
+
+                    if not b.at_sea:
+                        self.response.set_status(200)
+                        self.response.write("Boat was already docked")
+
+                    else:
+                        # check for provided slip number:
+                        foundSlip = False       # true until proven false
+                        if boat_data.has_key("number"):
+                            num = int(boat_data['number'])
+                            q = Slip.query().filter(Slip.number == num)
+                            amount = q.count()
+                            if amount == 1:
+                                slips = q.fetch()
+                                for slip in slips:
+                                    slip.current_boat = b.id
+                                    slip.arrival_date = date
+                                    slipNum = slip.number
+                                    b.at_sea = False
+                                    b.put()
+                                    slip.put()
+                                    foundSlip = True
+
+                        # no providede slip number, will find an open slip for you:
+                        else:
+                            # query all slips, find one without a boat:
+                            q = Slip.query()
+                            allSlips = q.order(Slip.number)
+                            for slip in allSlips.fetch():
+                                if (slip.current_boat is None):
+                                    slip.current_boat = b.id
+                                    slip.arrival_date = date
+                                    slipNum = slip.number
+                                    b.at_sea = False
+                                    b.put()
+                                    slip.put()
+                                    foundSlip = True
+                                    break
+
+                        if foundSlip == False:
+                            self.response.set_status(403)
+                            self.response.write("There was no slip available for dock!")
+                        else:
+                            self.response.set_status(200)
+                            self.response.write("boat was docked successfully at slip number: " + str(slipNum))
+
+            # boat with given id does not exist:
+            else:
+                self.response.set_status(404)
+                self.response.write("The provided boat id could not be found")
+
+        # no id provided in URL:
+        else:
+            self.response.set_status(404)
+            self.response.write("You must provide a boat with an id")
+
+
+
+
+
 
 
 
@@ -145,7 +272,6 @@ class SlipHandler(webapp2.RequestHandler):
 
         # create slip:
         new_slip = Slip()
-        new_slip.name = slip_data['name']
 
         # generate unused number for slip:
         q = Slip.query()
@@ -175,14 +301,24 @@ class SlipHandler(webapp2.RequestHandler):
         self.response.write("Here's the slip: " + json.dumps(slip_dict))
 
 
-
     # GET an existing SLIP
     def get(self, id=None):
         if id:
-            s = ndb.Key(urlsafe=id).get()
-            s_d = s.to_dict()
-            s_d['self'] = "/slip/" + id
-            self.response.write(json.dumps(s_d))
+            idInt = int(id)
+            s = Slip.get_by_id(idInt)
+            if (s != None):
+                s_d = s.to_dict()
+                s_d['self'] = "/slip/" + id
+
+                if (s.current_boat != None):
+                    s_d['current_boat_link'] = "/boat/" + s.current_boat
+
+                self.response.headers["Content-Type"] = "application/json"
+                self.response.write(json.dumps(s_d))
+
+            else:
+                self.response.set_status(404)
+                self.response.write("The slip with the given id could not be found")
 
         # display ALL slips:
         else:
@@ -193,8 +329,12 @@ class SlipHandler(webapp2.RequestHandler):
             # put all slips in a list for nice printing:
             slip_list = []
             for slip in allSlips:
-                slip = slip.to_dict()
-                slip_list.append(slip)
+                slipID = slip.id
+                slip_d = slip.to_dict()
+                slip_d['self'] = "/slip/" + slipID
+                if (slip.current_boat != None):
+                    slip_d['current_boat_link'] = "/boat/" + slip.current_boat
+                slip_list.append(slip_d)
             self.response.headers["Content-Type"] = "application/json"
             self.response.write(json.dumps(slip_list))
 
@@ -210,9 +350,21 @@ class SlipHandler(webapp2.RequestHandler):
                 self.response.write("Slip with given id could not be found")
 
             else:
-                s.key.delete()
-                self.response.set_status(200)
-                self.response.write("Slip with given id was removed successfully")
+                # remove boat from slip it's docked at
+                if (s.current_boat != None):
+                    bID = int(s.current_boat)
+                    b = Boat.get_by_id(bID)
+
+                    b.at_sea = True
+                    b.put()
+
+                    s.key.delete()
+                    self.response.set_status(200)
+                    self.response.write("Slip with given id was removed successfully")
+
+                else:
+                    self.response.write("No boat attached!")
+
 
         # no id supplied
         else:
@@ -220,6 +372,23 @@ class SlipHandler(webapp2.RequestHandler):
             self.response.write("You must supply an id for the slip to be removed")
 
 
+    # Modify a slip
+    def patch(self, id=None):
+        if id:
+                # look up slip with given id
+                s = Slip.get_by_id(int(id))
+
+                # TODO don't actually do this, debug
+                newNumber = s.number + 10
+                s.number = newNumber
+                s.put()
+
+                self.response.set_status(200)
+                self.response.write("I've updated your slip!")
+
+        else:
+            self.response.set_status(404)
+            self.response.write("You must provide a slip with an id")
 
 
 
@@ -229,25 +398,46 @@ class SlipHandler(webapp2.RequestHandler):
 
 
 
-
-
-
+# +--------------------------------------+
+# |          Main Page (Home)            |
+# +--------------------------------------+
 
 class MainPage(webapp2.RequestHandler):
     def get(self):
         self.response.headers['Content-Type'] = 'text/plain'
-        self.response.write('Hello, World 2!')
+        self.response.write(  'CS 496-400, Spring 2018: RESTful implementation\n'
+                            + '\n'
+                            + 'Site usage:\n'
+                            + 'GET      "/"                 - Home Directory (here)\n'
+                            + 'GET      "/boat"             - view all boats\n'
+                            + 'GET      "/boat/:boat_id"    - view a specific boat\n'
+                            + 'GET      "/slip"             - view all slips\n'
+                            + 'GET      "/slip/:slip_id"    - view a specific slip\n'
+                            + '\n'
+                            + 'POST     "/boat/"            - Create a new boat with given name, type, and length (all required)\n'
+                            + 'POST     "/slip/"            - Create a new slip\n'
+                            + 'DELETE   "/boat/:boat_id"    - delete a specific boat\n'
+                            + 'DELETE   "/slip/:slip_id"    - delete a specific slip\n'
+                            + 'PATCH    "/boat/:boat_id"    - dock/undock a specific boat (docking done manually or automatically)\n'
+                        )
 
 
 
+
+
+
+
+
+
+# +--------------------------------------+
+# |              Create App              |
+# +--------------------------------------+
 
 allowed_methods = webapp2.WSGIApplication.allowed_methods
 new_allowed_methods = allowed_methods.union(('PATCH',))
 webapp2.WSGIApplication.allowed_methods = new_allowed_methods
 app = webapp2.WSGIApplication([
     ('/', MainPage),
-    ('/fish', FishHandler),
-    ('/fish/(.*)', FishHandler),
     ('/boat', BoatHandler),
     ('/boat/(.*)', BoatHandler),
     ('/slip', SlipHandler),
